requirements.txt
PyQt5==5.15.9
pytesseract==0.3.10
Pillow==10.0.0
pyperclip==1.8.2
keyboard==0.13.5
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0